Re-added test base class, misc refactorings, including stop relying on event counts
