+++ src/test/groovy/main/MiscTests.groovy
@@ -11,7 +11,6 @@ import spock.util.concurrent.PollingConditions
 import util.TestUtil
 
 
-
 class MiscTests extends MikeAndConquerTestBase {
 
     enum GameSpeed
@@ -26,7 +25,6 @@ class MiscTests extends MikeAndConquerTestBase {
         Fastest
 
     }
-
 
 
     def setup() {
@@ -116,21 +114,21 @@ class MiscTests extends MikeAndConquerTestBase {
 
         where:
         unitType    | expectedTotalEvents   | expectedTimeInMillis  | gameSpeed
-//        "Jeep"      |  140                  |  30236                | "Slowest"
-//        "Jeep"      |  140                  |  15120                | "Slower"
-//        "Jeep"      |  140                  |  10082                | "Slow"
-//        "Jeep"      |  140                  |  7560                 | "Moderate"
-//        "Jeep"      |  140                  |  5040                 | "Normal"
-//        "Jeep"      |  140                  | 3697                  | "Fast"
-//        "Jeep"      |  140                  | 3024                  | "Faster"
-//        "Jeep"      |  140                  | 2855                  | "Fastest"
-//        "MCV"       |  302                  | 75597                 | "Slowest"
-//        "MCV"       |  302                  | 37801                 | "Slower"
-//        "MCV"       |  302                  | 25201                 | "Slow"
-//        "MCV"       |  302                  | 18900                 | "Moderate"
-//        "MCV"       |  302                  | 12601                 | "Normal"
-//        "MCV"       |  302                  | 9240                  | "Fast"
-//        "MCV"       |  302                  | 7560                  | "Faster"
+        "Jeep"      |  140                  |  30236                | "Slowest"
+        "Jeep"      |  140                  |  15120                | "Slower"
+        "Jeep"      |  140                  |  10082                | "Slow"
+        "Jeep"      |  140                  |  7560                 | "Moderate"
+        "Jeep"      |  140                  |  5040                 | "Normal"
+        "Jeep"      |  140                  | 3697                  | "Fast"
+        "Jeep"      |  140                  | 3024                  | "Faster"
+        "Jeep"      |  140                  | 2855                  | "Fastest"
+        "MCV"       |  302                  | 75597                 | "Slowest"
+        "MCV"       |  302                  | 37801                 | "Slower"
+        "MCV"       |  302                  | 25201                 | "Slow"
+        "MCV"       |  302                  | 18900                 | "Moderate"
+        "MCV"       |  302                  | 12601                 | "Normal"
+        "MCV"       |  302                  | 9240                  | "Fast"
+        "MCV"       |  302                  | 7560                  | "Faster"
         "MCV"       |  302                  | 7139                  | "Fastest"
     }
 
@@ -408,16 +406,16 @@ class MiscTests extends MikeAndConquerTestBase {
 
     }
 
-    Unit parseUnitFromEventData(String unitCreatedEventData) {
-        def unitDataObject = jsonSlurper.parseText(unitCreatedEventData)
-        Unit createdUnit = new Unit()
-        createdUnit.unitId = unitDataObject.UnitId
-        createdUnit.x = unitDataObject.X
-        createdUnit.y = unitDataObject.Y
-
-        return createdUnit
-
-    }
+//    Unit parseUnitFromEventData(String unitCreatedEventData) {
+//        def unitDataObject = jsonSlurper.parseText(unitCreatedEventData)
+//        Unit createdUnit = new Unit()
+//        createdUnit.unitId = unitDataObject.UnitId
+//        createdUnit.x = unitDataObject.X
+//        createdUnit.y = unitDataObject.Y
+//
+//        return createdUnit
+//
+//    }
